feat: leetcode request

+++ src/model/render.rs
@@ -15,7 +15,7 @@ pub struct Submission {
 pub struct User {
     pub username: String,
     pub country: String,
-    pub finish_time: u32,
+    pub finish_time: i64,
     pub global_rank: u32,
     pub score: u32,
     pub local_rank: u32,
