+++ src/model/website.rs
@@ -3,7 +3,7 @@ use super::render::{Contest, Submission};
 pub struct WebsiteUser {
     pub username: String,
     pub country: String,
-    pub finish_time: u32,
+    pub finish_time: i64,
     pub global_rank: u32,
     pub score: u32,
